Fixes a problem with pruning the system linking for linking scoring.
Suppose we have two responses X and Y which have the same TRFR A, but appear in different event frames in the system linking.  Suppose X is incorrectly assessed and Y is correctly assessed. If X scores higher than Y, by the previous behavior neither X nor Y would appears in the system linking for scoring. If Y scores higher than X, both would.  This commit fixes it so that the appearance of X or Y in the linking depends on its own assessment.

This also now applies the keepBest process within the scorer itself.