Handle duplicate responses.
For the TAC 2016 Events Track, the guidelines do not specify response or linking uniqueness.
For the 2015 eval, collapsing duplicate responses and response sets was a valid operation,
but isn't for the 2016 eval, as the CorpusLinking depends on the linking ids.

This commit adds handling for remapping duplicated linking/event hopper/doc event frame ids
and updating the CorpusLinking accordingly.